+++ bfe_config/bfe_conf/bfe_config_load.go
@@ -34,6 +34,7 @@ type BfeConfig struct {
 
 func SetDefaultConf(conf *BfeConfig) {
 	conf.Server.SetDefaultConf()
+	conf.HttpsBasic.SetDefaultConf()
 	conf.SessionCache.SetDefaultConf()
 	conf.SessionTicket.SetDefaultConf()
 }