+++ bfe_config/bfe_conf/conf_session_ticket.go
@@ -36,6 +36,10 @@ func (cfg *ConfigSessionTicket) SetDefaultConf() {
 }
 
 func (cfg *ConfigSessionTicket) Check(confRoot string) error {
+	if cfg.SessionTicketsDisabled {
+		return nil
+	}
+
 	return ConfSessionTicketCheck(cfg, confRoot)
 }