+++ CHANGELOG.md
@@ -10,6 +10,30 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v1.3.0] - 2021-09-16
+
+### Added
+- Support basic route rules that are similar to k8s ingress rules
+- Documents optimization
+
+### Changed
+- Ignore GREASE values for JA3 fingerprint
+
+
+## [v1.2.0] - 2021-06-21
+
+### Added
+- Set GOMAXPROCS according to real CPU quota if running on container 
+- Support condition primitive bfe_time_range/bfe_periodic_time_range
+- Support URI hash strategy
+- mod_block: support global request rules and ALLOW action
+- mod_header: support variables about client certificate
+- mod_access: support log to single file and stdout
+
+### Changed
+- Change primitive params of req_context_value_in
+
+
 ## [v1.1.0] - 2021-04-08
 
 ### Added
@@ -238,6 +262,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Flexible plugin framework to extend functionality. Based on the framework, developer can add new features rapidly
 - Detailed built-in metrics available for service status monitor
 
+[v1.3.0]: https://github.com/bfenetworks/bfe/compare/v1.2.0...v1.3.0
+[v1.2.0]: https://github.com/bfenetworks/bfe/compare/v1.1.0...v1.2.0
 [v1.1.0]: https://github.com/bfenetworks/bfe/compare/v1.0.0...v1.1.0
 [v1.0.0]: https://github.com/bfenetworks/bfe/compare/v0.12.0...v1.0.0
 [v0.12.0]: https://github.com/bfenetworks/bfe/compare/v0.11.0...v0.12.0